fix merge problem

+++ src/vmm/timer.rs
@@ -59,7 +59,7 @@ where
     let mut timers = timer_list.lock();
     let token = TOKEN.fetch_add(1, Ordering::Release);
     let event = VmmTimerEvent::new(token, handler);
-    timers.set(TimeValue::from_nanos(deadline), event);
+    timers.set(TimeValue::from_nanos(deadline as u64), event);
     token
 }
